+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
 - Console command to generate thumbs without job/queues
+- Config option for queue name for all mass updates (e.g. thumbs generating)
 ### Fixed
 - If resourceRevision was a clousure, an error occured
 - Fix AssetManager by adding S3 functions as trait